Fixed compiler warnings in vc1_dec and updated CMakeLists in vc1_common and scene_analyzer
Depends-on: I20d2aa32a9c820eb0c738eb68140e4f684cf0353

+++ _studio/shared/umc/codec/vc1_dec/src/umc_vc1_deblock.cpp
@@ -4,7 +4,7 @@
 //     This software is supplied under the terms of a license agreement or
 //     nondisclosure agreement with Intel Corporation and may not be copied
 //     or disclosed except in accordance with the terms of that agreement.
-//          Copyright(c) 2004-2013 Intel Corporation. All Rights Reserved.
+//          Copyright(c) 2004-2016 Intel Corporation. All Rights Reserved.
 //
 //
 //          VC-1 (VC1) decoder, deblocking
@@ -407,7 +407,7 @@ void Deblocking_ProgressivePpicture(VC1Context* pContext)
     LFilterType type_current = type;
     LFilterType type_current_next = type;
 
-    if (VC1_PROFILE_MAIN == type)
+    if (VC1_PROFILE_MAIN == static_cast<int>(type))
     {
         if ((pContext->m_MBs[0].mbType == VC1_MB_INTRA) || (pContext->m_MBs[0].m_pBlocks[0].blkType > VC1_BLK_INTER))
         {
